Remove subprocess as maybe tests are flaky on CI with it?
